Server side rendering

+++ src/utils/Storage.js
@@ -3,6 +3,9 @@
  */
 export default class Storage {
   static getItem(key, defaultValue) {
+    if (__SERVER__) {
+      return defaultValue;
+    }
     const itemStr = localStorage.getItem(key);
     if (!itemStr) {
       return defaultValue;
@@ -12,6 +15,9 @@ export default class Storage {
   }
 
   static setItem(key, item) {
+    if (__SERVER__) {
+      return;
+    }
     const itemStr = JSON.stringify({
       data: item,
     });
@@ -19,10 +25,16 @@ export default class Storage {
   }
 
   static removeItem(key) {
+    if (__SERVER__) {
+      return;
+    }
     localStorage.removeItem(key);
   }
 
   static clear() {
+    if (__SERVER__) {
+      return;
+    }
     localStorage.clear();
   }
 }
